refactor(p2): extract VirtualItems into sibling page (stage 3 of 4)
Moves the virtualization-pipeline section from ChatMessage.kt L324-L807
(in trimmed Stage 2 coord) into new sibling ChatMessageVirtualItems.kt
(same package): 3 consts/regex + sealed interface ChatMessageVirtualItem
(7 nested data classes/objects) + chatMessageVirtualizationPrewarmTexts +
buildChatMessageVirtualItems + shouldVirtualizeMarkdownContent +
ChatMessageVirtualItemContent + ChatMessageVirtualFooter.

Note: pre-cut #2 decl grep missed `internal sealed interface
ChatMessageVirtualItem` (L328-391) — pattern did not match `internal
sealed`. Caught during stage 3 execution; type travels with its only
producer/consumer (buildChatMessageVirtualItems + CMVI Content).

Byte-equal: bodies and signatures unchanged. All sibling-side decls
keep their pre-existing visibility (no widening inside sibling).

Visibility (coord-side widening for cross-file sibling access):
- hasRenderableChatMessageContent: private -> internal (CMVI Content uses)
- animateContentSizeIf: private -> internal (CMVF uses)
- rememberChatMessageTextStyle: private -> internal (CMVI Content uses)
- MessagePartsBlock: private -> internal (CMVI Content uses; will move
  to Stage 4 sibling, stays internal)

Coordinator strip: 1219 -> 729 (-490 with import cleanup).
Imports trimmed: 6 orphans (MarkdownParseResult, cachedMarkdownParseResult,
canRenderByTopLevelBlocks, parseMarkdownContent, topLevelBlockCount,
topLevelBlockKey).
Pre-existing dead preserved: getValue, setValue, fastForEach, height
(Modifier-extension lesson #4: grep-pseudo-alive but 0 real Modifier
usage).

Dead extracted-as-is (byte-equal, not deleted, playbook §1 #4):
chatMessageVirtualizationPrewarmTexts (0 callers anywhere),
ChatMessageVirtualMarkdownMinChars (internal, 0 cross-file callers),
ChatMessageVirtualMarkdownMinTableLines (private, only shouldVirtualize).

Baseline 3/467 unit tests preserved.
